fix bigmem offset calculation for qxl
when using big mem the 32bits offset to the rambase is just not enougth
move it into 64bits.

Thanks.

>From 32ea292c0c35640f325c6bf00228ef5661640c4c Mon Sep 17 00:00:00 2001
From: Izik Eidus <ieidus@redhat.com>
Date: Tue, 2 Jun 2009 13:25:52 -0400
Subject: [PATCH] spice: kvm-userpsace: fix bigmem offset calculation

when using big mem the 32bits offset to the rambase is just not enougth
move it into 64bits.

Thanks.

Message-ID: <4A2510D3.4040906@redhat.com>
RH-Upstream-status: not-applicable
Acked-by: Dor Laor <dlaor@redhat.com>
Acked-by: Yaniv Kamay <ykamay@redhat.com>
Acked-by: Arnon Gilboa <agilboa@redhat.com>
Bugzilla: 504237

+++ qemu/hw/qxl.c
@@ -90,12 +90,12 @@ typedef struct QXLState {
     uint32_t io_base;
     QXLRom *rom;
     QXLModes *modes;
-    uint32_t rom_offset;
+    uint64_t rom_offset;
     uint32_t rom_size;
 
     uint8_t *ram_start;
     QXLRam *ram;
-    uint32_t ram_offset;
+    uint64_t ram_offset;
     uint32_t ram_size;
     uint64_t ram_phys_addr;
 
@@ -145,7 +145,7 @@ struct PCIQXLDevice {
 typedef struct PCIVDIPortDevice {
     PCIDevice pci_dev;
     uint32_t io_base;
-    uint32_t ram_offset;
+    uint64_t ram_offset;
     uint32_t ram_size;
     VDIPortRam *ram;
     uint32_t connected;
